HS-284 Refactor following current structure
