+++ ui/src/components/Appliances/MinionsTable.vue
@@ -46,8 +46,8 @@
 <script setup lang="ts">
 import { useApplianceQueries } from '@/store/Queries/applianceQueries'
 import { useAppliancesStore } from '@/store/Views/appliancesStore'
-import { formatItemBgColor } from '@/helpers/formatting'
 import ChevronLeft from '@featherds/icon/navigation/ChevronLeft'
+import { formatItemBgColor } from './appliances.helpers'
 
 const appliancesStore = useAppliancesStore()
 const applianceQueries = useApplianceQueries()
